factor out spatial coverage display
